Move Demand related components to dedicated folder
