+++ packages/origin-ui/src/components/Demand/Demands.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
-import { PageContent } from '../elements/PageContent/PageContent';
+import { PageContent } from '../../elements/PageContent/PageContent';
 import { DemandTable } from './DemandTable';
 import { connect } from 'react-redux';
-import { getDemandsLink } from '../utils/routing';
-import { IStoreState } from '../types';
-import { getBaseURL } from '../features/selectors';
+import { getDemandsLink } from '../../utils/routing';
+import { IStoreState } from '../../types';
+import { getBaseURL } from '../../features/selectors';
 import { DemandForm } from './DemandForm';
 import { NavLink, Route, Redirect } from 'react-router-dom';
 import { DemandEdit } from './DemandEdit';
